+++ docs/src/background_task_lock_args_controller/background_task_lock_args_controller_curl3.bash
@@ -1,4 +1,4 @@
 sleep 1
-curl -i localhost:8000/hello-single?task_id=de01341028ac
+curl -i localhost:8000/hello-single?task_id=4ee301eb-6487-48a0-b6ed-e5f576accfc2
 echo && echo
-curl -i localhost:8000/hello-single?task_id=1d62d7c10b8b
+curl -i localhost:8000/hello-single?task_id=5ee301eb-6487-48a0-b6ed-e5f576accfc2